memory cache now has all data
